requirements.txt
Flask==1.1.2
line-bot-sdk==1.16.0
numpy==1.19.0
pillow==7.2.0
gunicorn
opencv-contrib-python
pandas